refactor(config): delete model-capabilities.ts; rename to in-tree-data (#2546 slice E)
Epic finale. After slices A-D, `model-capabilities.ts` was the last
direct importer of the legacy helper surface. This slice:

- Renames `model-capabilities.ts` → `in-tree-data.ts`. The file now
  contains only `DEFAULT_MODEL_CAPABILITIES` and `DEFAULT_MODEL_PER_CLI`
  (data-only). All helper functions (getModelCapabilities,
  findModelsByOutputModality, findModelsByInputModality,
  findModelsByToolCapability, findModelsByFeature, findModelsByProvider,
  findBestModelForOutput, modelSupportsAll, findModelsByCli) have been
  deleted from this file. Type and enum re-exports have moved to be
  read directly from `model-capabilities-types.ts`.

- Adds registry-backed equivalents of the still-needed helpers
  (findModelsByOutputModality, findModelsByInputModality,
  findModelsByToolCapability, findModelsByFeature, findModelsByProvider,
  findBestModelForOutput, modelSupportsAll) in
  `model-config-helpers.ts`. Same signatures, same return shape.

- Migrates the last two production consumers:
  - cli/capabilities-command.ts — modality consts now from
    model-capabilities-types.ts directly; find* helpers from
    model-config-helpers.ts.
  - mcp/tools/delegate-to-model-helpers.ts — modelSupportsAll from
    model-config-helpers.ts.

- Migrates 3 test files and the in-tree-entries.test.ts to read data
  from in-tree-data.ts.

- Updates config/index.ts re-exports — data from in-tree-data, types
  + enums + schemas from model-capabilities-types, registry-backed
  helpers from model-config-helpers.

- Deletes the legacy test files (model-capabilities.test.ts and
  model-capabilities-find.test.ts) — the helpers they tested are
  now registry-backed and covered by model-config-helpers.test.ts.

- Updates the indexer symbol-extractor tests to use
  model-config-helpers.ts as the large-file fixture (it has similar
  shape: many exported functions, types, classes).

- Updates config/model-availability.ts (a small touch that was
  missed in earlier slices — imports DEFAULT_MODEL_PER_CLI from the
  new path).

This closes #2605 and the entire #2546 epic. Every public consumer
of model capabilities now reads through `ModelRegistry` (via
`getDefaultRegistry()` or the helper layer); the in-tree data is
just a tier-2 input to the registry.

Closes #2605.
Closes #2546.
